src/core/models/proprietary/cloud_training/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
docker>=6.1.0
pytorch>=2.1.0
transformers>=4.36.0
datasets>=2.15.0
accelerate>=0.25.0
wandb>=0.16.0
tqdm>=4.66.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pydantic>=2.5.0
loguru>=0.7.0 